updates github action for package names & versions
It turns out that using an override -p:PackageId or the version equivalent in dotnet pack, recursively overrides these variables for all projects built. As this project contains a number of projects this did not work and led to naming clashes. So the fix is to manually replace the PackageId and Version strings in the csproj files as required for each build and pack